feat(recall): move settings into Experimental behind a master toggle
Recall's settings page folds into Settings > Experimental behind a new
recallEnabled master switch (default on). When off, readRecallSettings()
reports every behavioral flag as false — brief injection, proactive
recall, auto-capture, agent writes, code-graph indexing, and the engine
all shut off — while stored sub-settings survive for re-enable. The
project actions menu drops its Recall entry when disabled, and the
standalone Recall nav item is gone (stored "memory" panel redirects to
Experimental).
