+++ src/components/views/GeneralSettingsPage.tsx
@@ -110,6 +110,7 @@ export function GeneralSettingsPage() {
     claudeUsageLimitsEnabled: settings?.claudeUsageLimitsEnabled ?? false,
     claudeUsageLimitsShowSession: settings?.claudeUsageLimitsShowSession ?? true,
     claudeUsageLimitsShowWeekly: settings?.claudeUsageLimitsShowWeekly ?? true,
+    recallEnabled: settings?.recallEnabled ?? true,
     recallAutoCaptureEnabled: settings?.recallAutoCaptureEnabled ?? true,
     recallEngineEnabled: settings?.recallEngineEnabled ?? true,
     recallEngineHarness: settings?.recallEngineHarness ?? "claude-code",
