+++ src/components/views/SettingsPanel.tsx
@@ -9,7 +9,6 @@ import { BetaSettingsPage } from "./BetaSettingsPage";
 import { DefaultsSettingsPage } from "./DefaultsSettingsPage";
 import { GeneralSettingsPage } from "./GeneralSettingsPage";
 import { KeybindingsPage } from "./KeybindingsPage";
-import { MemorySettingsPage } from "./MemorySettingsPage";
 import { TerminalSettingsPage } from "./TerminalSettingsPage";
 import { ThemeSettingsPage } from "./ThemeSettingsPage";
 import { TermsSettingsPage } from "./TermsSettingsPage";
@@ -26,7 +25,6 @@ export const SETTINGS_PANEL_IDS = [
   "terminal",
   "theme",
   "voice",
-  "memory",
   "beta",
   "keybindings",
   "terms",
@@ -37,6 +35,8 @@ type NavItem = { id: SettingsPanelId; label: string; icon: IconName };
 
 function normalizeStoredPanel(stored: string | null, fallback: SettingsPanelId): SettingsPanelId {
   if (stored === "sandbox") return "beta";
+  // Recall settings moved from their own "memory" page into Experimental.
+  if (stored === "memory") return "beta";
   if (stored && SETTINGS_PANEL_IDS.includes(stored as SettingsPanelId)) {
     return stored as SettingsPanelId;
   }
@@ -100,7 +100,6 @@ export function SettingsPanel({
     { id: "terminal", label: "Terminal", icon: "terminal" },
     { id: "theme", label: "Theme", icon: "sun" },
     { id: "voice", label: "Voice", icon: "play" },
-    { id: "memory", label: "Recall", icon: "sparkles" },
     { id: "keybindings", label: "Keybindings", icon: "settings" },
   ];
 
@@ -332,8 +331,6 @@ export function SettingsPanel({
             <ThemeSettingsPage />
           ) : activePanel === "voice" ? (
             <VoiceCommandsPage />
-          ) : activePanel === "memory" ? (
-            <MemorySettingsPage />
           ) : activePanel === "beta" ? (
             <BetaSettingsPage />
           ) : activePanel === "keybindings" ? (
