+++ src/components/views/TerminalSettingsPage.tsx
@@ -54,6 +54,7 @@ export function TerminalSettingsPage() {
     claudeUsageLimitsEnabled: settings?.claudeUsageLimitsEnabled ?? false,
     claudeUsageLimitsShowSession: settings?.claudeUsageLimitsShowSession ?? true,
     claudeUsageLimitsShowWeekly: settings?.claudeUsageLimitsShowWeekly ?? true,
+    recallEnabled: settings?.recallEnabled ?? true,
     recallAutoCaptureEnabled: settings?.recallAutoCaptureEnabled ?? true,
     recallEngineEnabled: settings?.recallEngineEnabled ?? true,
     recallEngineHarness: settings?.recallEngineHarness ?? "claude-code",
