+++ src/components/views/ThemeSettingsPage.tsx
@@ -62,6 +62,7 @@ export function ThemeSettingsPage() {
     claudeUsageLimitsEnabled: settings?.claudeUsageLimitsEnabled ?? false,
     claudeUsageLimitsShowSession: settings?.claudeUsageLimitsShowSession ?? true,
     claudeUsageLimitsShowWeekly: settings?.claudeUsageLimitsShowWeekly ?? true,
+    recallEnabled: settings?.recallEnabled ?? true,
     recallAutoCaptureEnabled: settings?.recallAutoCaptureEnabled ?? true,
     recallEngineEnabled: settings?.recallEngineEnabled ?? true,
     recallEngineHarness: settings?.recallEngineHarness ?? "claude-code",
